Initial Quarkus Extension (#113)
* Initial Quarkus Extension

This creates a base quarkus extension for integrating with DJL along with an
example using the extension.

* Quarkus inject multiple Predictors

For a model, the predictor should only be used in a single thread. This change
only builds the model through quarkus, but the predictors are created during
dependency injection.

* Quarkus DJL runtime initialization with classes

* Support native build

It now points to the master version of quarkus because the
RuntimeInitializedPackageBuildItem has not been put into an official release.